% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{apply_by}
\alias{apply_by}
\title{Apply by dimension.}
\usage{
apply_by(X, MARGIN, FUN, na.exclude = TRUE, ..., simplify = TRUE)
}
\arguments{
\item{X}{A stratified feature table object. Preferrably of the stratifiedFeatureTable class.}

\item{MARGIN}{An single integer indicating which dimension to apply a given function on. 1 indicates by sample, 2 by feature and 3 by subtype.}

\item{FUN}{The function to be applied: see `?apply`.}

\item{na.exclude}{A boolean. Recommended to keep at TRUE.}

\item{...}{Optional arguments to FUN.}

\item{simplify}{A logical indicating whether results should be simplified if possible.}
}
\value{
A formatted data.frame.
}
\description{
A method to apply a function by dimension for stratified feature tables. Wraps around apply.
}
