% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifiedFeatureTable.R
\name{as.stratifiedFeatureTable}
\alias{as.stratifiedFeatureTable}
\alias{as.sft}
\title{Generate a stratified feature table object}
\usage{
as.stratifiedFeatureTable(x, sep = "\\\\|")

as.sft(x, sep = "\\\\|")
}
\arguments{
\item{x}{A table to be formatted to `stratifiedFeatureTable` format.}

\item{sep}{A character string to separate the terms.}
}
\description{
`as.stratifiedFeatureTable()` wrangles an existing table, such as a data frame, 
into the stratifiedFeatureTable format.
}
\examples{
sdf <- dummy_strat(10, 20, 10)

sft <- as.stratifiedFeatureTable(sdf)
}
