% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{board_argo}
\alias{board_argo}
\title{Make an argonautica themed dummy stratified data.}
\usage{
board_argo(
  nsubtypes = 96,
  nfeatures = 34,
  nsamples = 10,
  type = "by_subtype",
  p_missing = 1/2
)
}
\arguments{
\item{nsubtypes}{An integer. How many heroes}

\item{nfeatures}{An integer. How many skills}

\item{nsamples}{An integer. How many argos}

\item{type}{A character vector. One ofc("seq", "by_subtype").}

\item{p_missing}{A proportion (0-1). Indicates the number of feature-subtype pairs should be dropped.}
}
\value{
An argonautica themed formatted data.frame.
}
\description{
Make an argonautica themed normal dummy stratified feature data.
}
