% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dummy_names}
\alias{dummy_names}
\title{Provide dummy names data.}
\usage{
dummy_names(subject, n)
}
\arguments{
\item{subject}{A character vector. Which theme to provide.}

\item{n}{An integer. how many names to provide.}
}
\value{
A character vector of names.
}
\description{
Make dummy names for dummy_strat
}
