% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dummy_strat}
\alias{dummy_strat}
\title{Make dummy stratified data.}
\usage{
dummy_strat(nmicrobes, ngenes, nsamples)
}
\arguments{
\item{nmicrobes}{An integer. How many microbes}

\item{ngenes}{An integer. How many genes}

\item{nsamples}{An integer. How many samples}
}
\value{
A formatted data.frame.
}
\description{
Make normal dummy stratified feature data.
}
