% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getFeature}
\alias{getFeature}
\title{Extract all subtypes of a feature.}
\usage{
getFeature(x, f)
}
\arguments{
\item{x}{A stratified feature table object. Preferably of the stratifiedFeatureTable class.}

\item{f}{Character or integer. Name or index of feature to extract.}
}
\value{
A matrix with samples as rows and feature subtypes as columns
}
\description{
Extract a matrix from a stratified feature table
}
