% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{stratifiedFeatureTable-class}
\alias{stratifiedFeatureTable-class}
\title{An S4 class stratifiedFeatureTable.}
\description{
This is a specialized S4 class to represent stratified feature data.
}
