% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{stratifiedFeatureTable}
\alias{stratifiedFeatureTable}
\alias{sft}
\title{Make a suitably shaped array into a stratifiedFeatureTable.}
\usage{
stratifiedFeatureTable(x)

sft(x)
}
\arguments{
\item{x}{Asuitably shaped array to be changed to `stratifiedFeatureTable` format.}
}
\value{
A stratifiedFeatureTable object.
}
\description{
Make a stratifiedFeatureTable, a specialized S4 class to represent stratified feature data.
}
